Removing references to deprecated 1.5 version
These notes are very old, and while we still support patch versions of Ops Metrics 1.4 & Ops Metrics 1.6, version 1.5 was deprecated due to instability and has not been an available release for a long time. Removing references to it as it is already 2+ versions behind current and could be confusing to be on this page when it is not a valid release option for PCF users.