Removing RN references to deprecated 1.5 version
These notes are very old, and while we still support patch versions of Ops Metrics 1.4 & Ops Metrics 1.6, version 1.5 was deprecated due to instability and has not been an available release for a long time. Removing references to it as it is already 2+ versions behind current and could be confusing to be on this page when it is not a valid release option for PCF users. I did leave one note in upgrade path just in case someone still has a bad 1.5 out there and wants to move to 1.6. I also updated the version number and required stemcell to latest available on PivNet. I expect to make another PR soon when we release another 1.6 patch in support of an existing CVE