Correction to KI based on loggregator patch work
Updating submitted RN jmx note. Loggregator team communicated that the metric being backpatched needed a slight name tweak. Additionally it wasn't as clean as "from this to that" as originally communicated so altering the reference that sounded like this former metric was a name-correct metric pre 1.10; it was not. Based on the work loggregator is implementing today to get into an ERT 1.10 patch, this corrected note should now be accurate.