PASW Smoke Tests + internetless + OCSP
Currently, PASW smoke tests are defaulting to off because if a customer has a particular environment configuration, that smoke test will fail. We would like customers to turn it on, but there is a workaround for customers with that particular configuration to make this work. If this explanation belongs in a Known Issues section instead of New Features, I am happy to move it!