Updating the status for 2.8.4 as Withdrawn
Ops Manager team got to know about a breaking change in BOSH DNS version which was shipped with Ops Manager 2.8.4 last week. This causes failures in the integration of PAS Windows.
We have made the 2.8.4 release private (Admins only) on PivNet. We have also sent an alert to those who downloaded that release. 

We wanted to update the 2.8.4 release notes to clarify that the release has been pulled/withdrawn with a note: “This release has been removed from Pivotal Network because it shipped with an unintentional breaking change in BOSH-DNS which causes PAS Windows deployments to fail.”